Refactored all <reference> tags to import statements
