+++ src/Cookie.ts
@@ -1,0 +1,9 @@
+export interface ICookieOption {
+    domain?: string;
+    path?: string;
+    secure?: boolean;
+    httpOnly?: boolean;
+    expires?: Date;
+    maxAge?: number;
+    signed?: boolean;
+}
