--- src/TSFramework.ts
@@ -1,16 +1,0 @@
-/// <reference path="../typings/tsd.d.ts" />
-/// <reference path="../typings/waterline.d.ts" />
-
-/// <reference path="Global.ts" />
-/// <reference path="Express.ts" />
-/// <reference path="Declaration.ts" />
-/// <reference path="Model.ts" />
-/// <reference path="Configuration.ts" />
-/// <reference path="Request.ts" />
-/// <reference path="Response.ts" />
-/// <reference path="Reply.ts" />
-/// <reference path="Controller.ts" />
-/// <reference path="Router.ts" />
-/// <reference path="Application.ts" />
-
-module.exports = TS;
